+++ src/vue/apps/resume/Resume.vue
@@ -244,8 +244,10 @@
                         //             }
                         //         ]
                         //     }
-                    }
-                }
+                    },
+                    pusheenStatus:'dancing',
+                    pusheenRememberMe: false,
+                },
             }
         }
     }